(chore) removed registration for new body from outside
