+++ packages/esm-morgue-app/src/tabs/tabs.scss
@@ -3,203 +3,6 @@
 @use '@carbon/colors';
 @use '~@openmrs/esm-styleguide/src/vars' as *;
 
-.container {
-  margin: 2rem 0;
-}
-
-.emptyStateContainer,
-.loaderContainer {
-  @extend .container;
-}
-
-.serviceContainer {
-  background-color: $ui-02;
-  border: 1px solid $ui-03;
-  width: 100%;
-  margin: 0 auto;
-  max-width: 95vw;
-  padding-bottom: 0;
-
-  :has(.filterEmptyState) {
-    border-bottom: none;
-  }
-}
-
-.left-justified-items {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  cursor: pointer;
-  align-items: center;
-}
-
-.filterContainer {
-  flex: 1;
-
-  :global(.cds--dropdown__wrapper--inline) {
-    gap: 0;
-  }
-
-  :global(.cds--list-box__menu-icon) {
-    height: 1rem;
-  }
-
-  :global(.cds--list-box__menu) {
-    min-width: max-content;
-  }
-
-  :global(.cds--list-box) {
-    margin-left: layout.$spacing-03;
-  }
-}
-
-.menu {
-  margin-left: layout.$spacing-03;
-}
-
-.headerContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: layout.$spacing-04 layout.$spacing-05;
-  background-color: $ui-02;
-}
-
-.backgroundDataFetchingIndicator {
-  align-items: center;
-  display: flex;
-  flex: 1;
-  justify-content: space-between;
-
-  &:global(.cds--inline-loading) {
-    max-height: layout.$spacing-05;
-  }
-}
-
-.tableContainer section {
-  position: relative;
-}
-
-.tableContainer a {
-  text-decoration: none;
-}
-
-.pagination {
-  overflow: hidden;
-
-  &:global(.cds--pagination) {
-    border-top: none;
-  }
-}
-
-.hiddenRow {
-  display: none;
-}
-
-.emptyRow {
-  padding: 0 1rem;
-  display: flex;
-  align-items: center;
-}
-
-.visitSummaryContainer {
-  width: 100%;
-  max-width: 768px;
-  margin: 1rem auto;
-}
-
-.expandedActiveVisitRow > td > div {
-  max-height: max-content !important;
-}
-
-.expandedActiveVisitRow td {
-  padding: 0 2rem;
-}
-
-.expandedActiveVisitRow th[colspan] td[colspan] > div:first-child {
-  padding: 0 1rem;
-}
-
-.action {
-  margin-bottom: layout.$spacing-03;
-}
-
-.illo {
-  margin-top: layout.$spacing-05;
-}
-
-.content {
-  @include type.type-style('heading-compact-01');
-  color: $text-02;
-  margin-top: layout.$spacing-05;
-  margin-bottom: layout.$spacing-03;
-}
-
-.desktopHeading,
-.tabletHeading {
-  text-align: left;
-  text-transform: capitalize;
-  flex: 1;
-
-  h4 {
-    @include type.type-style('heading-compact-02');
-    color: $text-02;
-
-    &:after {
-      content: '';
-      display: block;
-      width: 2rem;
-      padding-top: 3px;
-      border-bottom: 0.375rem solid;
-      @include brand-03(border-bottom-color);
-    }
-  }
-}
-
-.tile {
-  text-align: center;
-  border: 1px solid $ui-03;
-}
-
-.menuitem {
-  max-width: none;
-}
-
-.filterEmptyState {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: layout.$spacing-05;
-  margin: layout.$spacing-09;
-  text-align: center;
-}
-
-.filterEmptyStateTile {
-  margin: auto;
-}
-
-.filterEmptyStateContent {
-  @include type.type-style('heading-compact-02');
-  color: $text-02;
-  margin-bottom: 0.5rem;
-}
-
-.filterEmptyStateHelper {
-  @include type.type-style('body-compact-01');
-  color: $text-02;
-}
-
-.actionBtn {
-  display: flex;
-  gap: 0.5rem;
-  margin-right: layout.$spacing-05;
-}
-
-.mainSection {
-  display: grid;
-  grid-template-columns: 16rem 1fr;
-}
-
 .tabsContainer {
   display: flex;
   justify-content: space-between;
